+++ packages/react-workspace/nimble-react/src/theme-provider/index.ts
@@ -1,5 +1,6 @@
 import { ThemeProvider } from '@ni/nimble-components/dist/esm/theme-provider';
+import { Theme } from '@ni/nimble-components/dist/esm/theme-provider/types';
 import { wrap } from '../utilities/react-wrapper';
 
-export { type ThemeProvider };
+export { type ThemeProvider, Theme };
 export const NimbleThemeProvider = wrap(ThemeProvider);