+++ packages/react-workspace/nimble-react/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @ni/nimble-react
 
-<!-- This log was last generated on Wed, 17 Dec 2025 23:15:32 GMT and should not be manually modified. -->
+<!-- This log was last generated on Fri, 02 Jan 2026 06:16:54 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.9.0
+
+Fri, 02 Jan 2026 06:16:54 GMT
+
+### Minor changes
+
+- Switch to package exports ([ni/nimble@66f9a70](https://github.com/ni/nimble/commit/66f9a70c13c5a2f8605b86036f1faded63e2bf93))
+
+## 0.8.2
+
+Fri, 26 Dec 2025 05:48:06 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v34.10.3
+
+## 0.8.1
+
+Wed, 24 Dec 2025 23:16:44 GMT
+
+### Patches
+
+- Export Theme type from ThemeProvider ([ni/nimble@76a4daf](https://github.com/ni/nimble/commit/76a4daff934119d14006bdab4859fc05c21cf42d))
+
+## 0.8.0
+
+Wed, 24 Dec 2025 01:13:38 GMT
+
+### Minor changes
+
+- React 19 support ([ni/nimble@f02af57](https://github.com/ni/nimble/commit/f02af57ebb0bd171e3778b06bb05ad1d140317bf))
 
 ## 0.7.12